+++ packages/frontend/src/providers/Explorer/ExplorerProvider.tsx
@@ -88,6 +88,9 @@ const defaultState: ExplorerReduceState = {
         customDimension: {
             isOpen: false,
         },
+        additionalMetricWriteBack: {
+            isOpen: false,
+        },
     },
 };
 
@@ -792,6 +795,18 @@ function reducer(
                 },
             };
         }
+        case ActionType.TOGGLE_ADDITIONAL_METRIC_WRITE_BACK_MODAL: {
+            return {
+                ...state,
+                modals: {
+                    ...state.modals,
+                    additionalMetricWriteBack: {
+                        isOpen: !state.modals.additionalMetricWriteBack.isOpen,
+                        ...(action.payload && { ...action.payload }),
+                    },
+                },
+            };
+        }
         case ActionType.SET_COLUMN_ORDER: {
             return {
                 ...state,
@@ -1277,6 +1292,16 @@ const ExplorerProvider: FC<
         [],
     );
 
+    const toggleAdditionalMetricWriteBackModal = useCallback(
+        (args?: { item?: AdditionalMetric }) => {
+            dispatch({
+                type: ActionType.TOGGLE_ADDITIONAL_METRIC_WRITE_BACK_MODAL,
+                payload: args,
+            });
+        },
+        [],
+    );
+
     const setColumnOrder = useCallback((order: string[]) => {
         dispatch({
             type: ActionType.SET_COLUMN_ORDER,
@@ -1565,6 +1590,7 @@ const ExplorerProvider: FC<
             editAdditionalMetric,
             removeAdditionalMetric,
             toggleAdditionalMetricModal,
+            toggleAdditionalMetricWriteBackModal,
             addTableCalculation,
             deleteTableCalculation,
             updateTableCalculation,
@@ -1615,6 +1641,7 @@ const ExplorerProvider: FC<
             toggleCustomDimensionModal,
             toggleFormatModal,
             updateMetricFormat,
+            toggleAdditionalMetricWriteBackModal,
             replaceFields,
         ],
     );